just one test

+++ ft_print_hex_upper.c
@@ -1,8 +1,18 @@
-
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_print_hex_upper.c                               :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: abouramt <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/12/02 14:20:00 by abouramt          #+#    #+#             */
+/*   Updated: 2023/12/02 14:20:09 by abouramt         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_print_hex_upper(int nb)
+int	ft_print_hex_upper(unsigned long nb)
 {
 	int		count;
 	char	*hex;
